pyproject.toml
[tool.poetry]
name = "ferbot"
version = "0.1.0"
description = ""
authors = ["Bernard Crnković <bernard.crnkovic@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
selenium = "^4.5.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
attrs = "^22.1.0"
python-dotenv = "^0.21.0"

[tool.poetry.scripts]
materijali = "ferbot.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
